+++ packages/core/src/connectors/mock.test.ts
@@ -30,7 +30,7 @@ test('behavior: features.connectError', () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to connect.
-    Version: viem@2.23.5]
+    Version: viem@2.23.12]
   `)
 })
 
@@ -58,7 +58,7 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to sign typed data.
-    Version: viem@2.23.5]
+    Version: viem@2.23.12]
   `)
 
   expect(
@@ -70,7 +70,7 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to switch chain.
-    Version: viem@2.23.5]
+    Version: viem@2.23.12]
   `)
 
   expect(
@@ -82,7 +82,7 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to switch chain.
-    Version: viem@2.23.5]
+    Version: viem@2.23.12]
   `)
 
   expect(
@@ -94,7 +94,7 @@ test('behavior: connector.getProvider request errors', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to sign message.
-    Version: viem@2.23.5]
+    Version: viem@2.23.12]
   `)
 })